Rename `Subscriber` to `Action` #84
