+++ src/action/mod.rs
@@ -3,14 +3,14 @@ mod command;
 mod handler;
 mod publisher;
 mod routine;
-mod subscriber;
-mod subscribers;
+mod action;
+mod actions;
 mod types;
 
 pub use command::*;
 pub use handler::*;
 pub use publisher::*;
 pub use routine::*;
-pub use subscriber::*;
-pub use subscribers::*;
+pub use action::*;
+pub use actions::*;
 pub use types::*;
